Add padding for larger desktops
